tunnel: rename management-plane firewall + renumber spec to avoid VPN-broker collision
Upstream main independently shipped a per-VM VPN broker (spec/19-vpn-broker.md) +
per-VM firewall (spec/20-firewall.md) with files that collide with this feature's
management-plane lockdown. Rename to keep them distinct:
- scripts/firewall-{apply,revert,confirm}.py -> scripts/mgmt-firewall-*.py
- scripts/lib/atlas/firewall.py -> mgmt_firewall.py (+ test_mgmt_firewall.py)
- spec/19-tunnel.md -> spec/21-tunnel.md
and update all references (scripts_catalog, central_link, tests, spec cross-refs).
The nft table was already inet atlas_mgmt (distinct from the per-VM inet atlas).
